[#2] evaluating m-1-2 works. config.yaml simplified.
