+++ explore/explore_bart_logits_shape.py
@@ -1,6 +1,6 @@
 from transformers import BartTokenizer, BartForConditionalGeneration
 
-from data import IdiomifyDataModule
+from datamodules import IdiomifyDataModule
 
 
 CONFIG = {
